fix analytics

+++ src/main/java/com/ems/campaign/model/CampaignAnalytics.java
@@ -107,7 +107,7 @@ public class CampaignAnalytics {
         int emailOpenCount = getEmailOpenCountFromEmailDetails(emailDetailsList);
         int totalSubscribers = emailDetailsList.size();
 
-        return (emailOpenCount / totalSubscribers) * 100;
+        return ((double) emailOpenCount / totalSubscribers) * 100;
     }
 
     private int getEmailOpenCountFromEmailDetails(List<EmailDetails> emailDetailsList){
@@ -128,7 +128,7 @@ public class CampaignAnalytics {
                 linkClickCount++;
             }
         }
-        return (linkClickCount / subscriberCount) * 100;
+        return ((double) linkClickCount / subscriberCount) * 100;
     }
 
     private double calculateUnsubscribeRate(List<EmailDetails> emailDetailsList){
@@ -140,6 +140,6 @@ public class CampaignAnalytics {
                 unsubscribersCount++;
             }
         }
-        return (unsubscribersCount / subscriberCount) * 100;
+        return ((double)unsubscribersCount / subscriberCount) * 100;
     }
 }
